+++ packages/io-sign/src/event.ts
@@ -1,6 +1,11 @@
 import { IsoDateFromString } from "@pagopa/ts-commons/lib/dates";
 import * as t from "io-ts";
 import * as TE from "fp-ts/lib/TaskEither";
+import * as E from "fp-ts/lib/Either";
+import * as RTE from "fp-ts/lib/ReaderTaskEither";
+import * as L from "@pagopa/logger";
+
+import { pipe } from "fp-ts/lib/function";
 import { Id, newId } from "./id";
 import {
   SignatureRequestDraft,
@@ -53,6 +58,14 @@ const BaseEvent = t.type({
 
 type BaseEvent = t.TypeOf<typeof BaseEvent>;
 
+type SignatureRequest =
+  | SignatureRequestDraft
+  | SignatureRequestSigned
+  | SignatureRequestReady
+  | SignatureRequestToBeSigned
+  | SignatureRequestWaitForQtsp
+  | SignatureRequestRejected;
+
 export const BillingEvent = t.intersection([
   BaseEvent,
   t.type({ name: t.literal(EventName.SIGNATURE_SIGNED) }),
@@ -90,15 +103,7 @@ export type AnalyticsEvent = t.TypeOf<typeof AnalyticsEvent>;
 
 export const createAnalyticsEvent =
   (eventName: EventName) =>
-  (
-    signatureRequest:
-      | SignatureRequestDraft
-      | SignatureRequestSigned
-      | SignatureRequestReady
-      | SignatureRequestToBeSigned
-      | SignatureRequestWaitForQtsp
-      | SignatureRequestRejected
-  ): AnalyticsEvent => ({
+  (signatureRequest: SignatureRequest): AnalyticsEvent => ({
     id: newId(),
     name: eventName,
     signatureRequestId: signatureRequest.id,
@@ -120,11 +125,68 @@ export type SendEvent = (
 export type CreateAndSendAnalyticsEvent = (
   eventName: EventName
 ) => (
-  signatureRequest:
-    | SignatureRequestDraft
-    | SignatureRequestSigned
-    | SignatureRequestReady
-    | SignatureRequestToBeSigned
-    | SignatureRequestWaitForQtsp
-    | SignatureRequestRejected
+  signatureRequest: SignatureRequest
 ) => TE.TaskEither<Error, typeof signatureRequest>;
+
+type EventData = {
+  body: GenericEvent;
+};
+
+type EventDataBatch = {
+  tryAdd(eventData: EventData): boolean;
+};
+
+type EventProducerClient = {
+  createBatch(): Promise<EventDataBatch>;
+  close: () => Promise<void>;
+  sendBatch(batch: EventDataBatch): Promise<void>;
+};
+
+type EventAnalyticsClient = {
+  eventAnalyticsClient: EventProducerClient;
+};
+
+export const sendEvent =
+  (
+    event: GenericEvent
+  ): RTE.ReaderTaskEither<EventAnalyticsClient, Error, GenericEvent> =>
+  ({ eventAnalyticsClient }) =>
+    pipe(
+      TE.tryCatch(() => eventAnalyticsClient.createBatch(), E.toError),
+      TE.chain((eventDataBatch) =>
+        eventDataBatch.tryAdd({ body: event })
+          ? TE.right(eventDataBatch)
+          : TE.left(new Error("Unable to add new events to event batch!"))
+      ),
+      TE.chain((eventDataBatch) =>
+        TE.tryCatch(
+          () => eventAnalyticsClient.sendBatch(eventDataBatch),
+          E.toError
+        )
+      ),
+      TE.map(() => event)
+    );
+
+export const createAndSendAnalyticsEvent =
+  (signatureRequest: SignatureRequest) => (eventName: EventName) =>
+    pipe(
+      signatureRequest,
+      createAnalyticsEvent(eventName),
+      sendEvent,
+      RTE.map(() => signatureRequest),
+      RTE.chainFirstW(() =>
+        L.debugRTE("Send analytics event", { eventName, signatureRequest })
+      ),
+      // This is a fire and forget operation
+      RTE.altW(() =>
+        pipe(
+          RTE.right(signatureRequest),
+          RTE.chainFirst(() =>
+            L.errorRTE("Unable to send analytics event", {
+              eventName,
+              signatureRequest,
+            })
+          )
+        )
+      )
+    );