+++ packages/io-sign/src/infra/azure/event-hubs/event.ts
@@ -1,16 +1,14 @@
 import { pipe } from "fp-ts/lib/function";
 
 import * as TE from "fp-ts/lib/TaskEither";
-import * as E from "fp-ts/lib/Either";
 import { EventHubProducerClient } from "@azure/event-hubs";
-import * as L from "@pagopa/logger";
-import { ConsoleLogger } from "../../console-logger";
 
 import {
-  createAnalyticsEvent,
+  createAndSendAnalyticsEvent,
   CreateAndSendAnalyticsEvent,
   EventName,
   GenericEvent,
+  sendEvent,
   SendEvent,
 } from "../../../event";
 import {
@@ -21,25 +19,15 @@ import {
   SignatureRequestWaitForQtsp,
   SignatureRequestRejected,
 } from "../../../signature-request";
+import { ConsoleLogger } from "../../console-logger";
 
 export const makeSendEvent =
-  (client: EventHubProducerClient): SendEvent =>
+  (eventAnalyticsClient: EventHubProducerClient): SendEvent =>
   (event: GenericEvent) =>
-    pipe(
-      TE.tryCatch(() => client.createBatch(), E.toError),
-      TE.chain((eventDataBatch) =>
-        eventDataBatch.tryAdd({ body: event })
-          ? TE.right(eventDataBatch)
-          : TE.left(new Error("Unable to add new events to event hub batch!"))
-      ),
-      TE.chain((eventDataBatch) =>
-        TE.tryCatch(() => client.sendBatch(eventDataBatch), E.toError)
-      ),
-      TE.map(() => event)
-    );
+    pipe({ eventAnalyticsClient }, sendEvent(event));
 
 export const makeCreateAndSendAnalyticsEvent =
-  (client: EventHubProducerClient): CreateAndSendAnalyticsEvent =>
+  (eventAnalyticsClient: EventHubProducerClient): CreateAndSendAnalyticsEvent =>
   (eventName: EventName) =>
   (
     signatureRequest:
@@ -51,30 +39,9 @@ export const makeCreateAndSendAnalyticsEvent =
       | SignatureRequestRejected
   ): TE.TaskEither<Error, typeof signatureRequest> =>
     pipe(
-      signatureRequest,
-      createAnalyticsEvent(eventName),
-      makeSendEvent(client),
-      TE.map(() => signatureRequest),
-      TE.chainFirstIOK(() =>
-        L.debug("Send analytics event", {
-          eventName,
-          signatureRequest,
-        })({
-          logger: ConsoleLogger,
-        })
-      ),
-      // This is a fire and forget operation
-      TE.altW(() =>
-        pipe(
-          TE.right(signatureRequest),
-          TE.chainFirstIOK(() =>
-            L.error("Unable to send analytics event", {
-              eventName,
-              signatureRequest,
-            })({
-              logger: ConsoleLogger,
-            })
-          )
-        )
-      )
+      {
+        eventAnalyticsClient,
+        logger: ConsoleLogger,
+      },
+      pipe(eventName, createAndSendAnalyticsEvent(signatureRequest))
     );